MDL-39913 assignsubmission_file: test of hierarchical checkboxes
Using CSS to indent the child types of a group, and to show or hide
the children when the group checkbox is selected. A form rule disables
the child checkboxes, and the disabled state is then tested by the CSS
to hide them from the page, thus 'collapsing' the group.